# Rename buffer to byteReader