This is a large change intended to reduce the surface area of the
byteReader (nee buffer) type. The original design came from iopipe where
the buffer type was an abstract implementation and the reader or scanner
built on top of it to extend the buffer then fill it from a reader. In
this implementation this separation of concerns didn't make sense as
there would only be one buffer impl, and it made more sense that the
byteReader knew how to refill itself, thereby removing several helper
methods like avail and remaining, bringing the design close to the
original iopipe API.

Overall performance is unaffected as extend only happens when the buffer
is exhausted, profiling shows this to be sub 1%.
```
benchstat {old,new}.txt
name                             old time/op    new time/op    delta
Scanner/canada.json.gz-16          3.60ms ± 2%    3.56ms ± 0%    ~     (p=0.690 n=5+5)
Scanner/citm_catalog.json.gz-16    1.63ms ± 1%    1.58ms ± 0%  -3.09%  (p=0.016 n=5+4)
Scanner/twitter.json.gz-16          837µs ± 1%     846µs ± 0%  +1.06%  (p=0.008 n=5+5)
Scanner/code.json.gz-16            3.69ms ± 1%    3.72ms ± 2%    ~     (p=0.421 n=5+5)
Scanner/example.json.gz-16         16.7µs ± 1%    16.7µs ± 1%    ~     (p=1.000 n=5+5)
Scanner/sample.json.gz-16           504µs ± 0%     477µs ± 0%  -5.39%  (p=0.008 n=5+5)

name                             old speed      new speed      delta
Scanner/canada.json.gz-16         625MB/s ± 2%   632MB/s ± 0%    ~     (p=0.690 n=5+5)
Scanner/citm_catalog.json.gz-16  1.06GB/s ± 1%  1.10GB/s ± 0%  +3.19%  (p=0.016 n=5+4)
Scanner/twitter.json.gz-16        755MB/s ± 1%   747MB/s ± 0%  -1.05%  (p=0.008 n=5+5)
Scanner/code.json.gz-16           526MB/s ± 1%   522MB/s ± 2%    ~     (p=0.421 n=5+5)
Scanner/example.json.gz-16        779MB/s ± 1%   779MB/s ± 1%    ~     (p=1.000 n=5+5)
Scanner/sample.json.gz-16        1.36GB/s ± 0%  1.44GB/s ± 0%  +5.70%  (p=0.008 n=5+5)

name                             old alloc/op   new alloc/op   delta
Scanner/canada.json.gz-16           0.00B          0.00B         ~     (all equal)
Scanner/citm_catalog.json.gz-16     0.00B          0.00B         ~     (all equal)
Scanner/twitter.json.gz-16          0.00B          0.00B         ~     (all equal)
Scanner/code.json.gz-16             0.00B          0.00B         ~     (all equal)
Scanner/example.json.gz-16          0.00B          0.00B         ~     (all equal)
Scanner/sample.json.gz-16           0.00B          0.00B         ~     (all equal)

name                             old allocs/op  new allocs/op  delta
Scanner/canada.json.gz-16            0.00           0.00         ~     (all equal)
Scanner/citm_catalog.json.gz-16      0.00           0.00         ~     (all equal)
Scanner/twitter.json.gz-16           0.00           0.00         ~     (all equal)
Scanner/code.json.gz-16              0.00           0.00         ~     (all equal)
Scanner/example.json.gz-16           0.00           0.00         ~     (all equal)
Scanner/sample.json.gz-16            0.00           0.00         ~     (all equal)
```